fs/binfmt_elf: Restrict string capabilities
This patch narrows down the permissions and bounds of initial
capabilities pointing to argument and environment variable strings
(elements of argv and envp).

This is a reduced version of Téo's patch ("fs: Handle exact bounds
for argv and envp"), which had to be reverted as it could not handle
empty strings. This version does not handle representability at all,
as a result string capabilities may overlap. To properly handle
string representability, it seems unavoidable to pass additional
metadata to binfmt_elf, to clarify if and where padding was added by
copy_strings().

Co-developed-by: Teo Couprie Diaz <teo.coupriediaz@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>