fs/binfmt_elf: Make AT_BASE an address again
AT_BASE was once used in purecap as root capability for the
interpreter. It was superseded by AT_CHERI_INTERP_*_CAP and no known
current software relies on it being a capability. Turn it back to a
plain address, consistently with the PCuABI specification (which
makes no special provision for AT_BASE, implying that it is a 64-bit
value).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>