bpf: jit: allocate stack in vmalloc region
arm64 JIT'd bpf programs on Morello currently re-use the existing kernel
stack in the kernel logical memory map (VMAP_STACK is not available in
PCuABI).

Since JIT programs will be running in a compartment with memory accesses
limited to the vmalloc region, new stacks for each eBPF program must be
created here to allow access and provide isolation of stacks.

Allocate and free new page sized/aligned stacks in the vmalloc region at
the same time as the binary image to ensure they last for the lifetime of
the program.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>