bpf: debug: disable prologue size check
We're about to add a lot of instructions to the prologue. To avoid
having to recalculate PROLOGUE_OFFSET every time and keep that up to
date, disable the static #define and check in build_prologue.

The check in build_prologue seems like it should know what the correct
offset is anyway, and the only place this offset is used is in
build_body which is called after build_prologue in bpf_int_jit_compile.

Set the offset in build_prologue to save manually calculating/updating
the #define.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>