fs/binfmt_elf: Create appropriate reservations in PCuABI
binfmt_elf already creates appropriate initial reservations thanks
to elf_map() creating an initial mapping with the total size of the
executable/interpreter. vm_mmap() will then automatically create a
reservation of the right size. However total_size is only computed for
dynamic executables at the moment, so for reservations to be correct
for static executables too we compute it in the static case as well.

Additionally, vm_brk_flags() is not aware of reservations, so we
shouldn't use it. An equivalent vm_mmap() call is made instead.

Signed-off-by: Amit Daniel Kachhap <amitdaniel.kachhap@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>