drm: Fix purecap vblank handling
Vblank is a bit different because it produces events that then can
copy pointers/capabilities back to userspace via a read on the DRM
device as opposed to the ioctl. The compat handler called the normal
native/purecap one and this happened to work for compat but not for
purecap. Make the shared "native" purecap handling work for both and
share that infra properly to avoid copy and paste.

Signed-off-by: Carsten Haitzler <Carsten.Haitzler@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>