drm: fix up purecap handling of all iotcls in libdrm test tools
The modetest, proptest vbltest tools from drm all work with this as the
ioctls they use with this patch now handle purecap ABI. The general
design is for these ioctls to have an import + export code "blob"
that converts to to/from in-kernel-data based on if it's a compat or
purecap caller and embedded compat structs together with the ioctl
handling. This is less invasive and less copy + paste based than
alternative solutions, thus making it more maintainable as a patchset.
There is no need to have special different ioctl number handling as
drm already handles this in generic infrastructure to make sure the
compat and native purecap ioctl numbers map to the same ioctl func
unless its explicitly in the 32bit (compat) handling.

Signed-off-by: Carsten Haitzler <Carsten.Haitzler@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>