arm64: morello: Ensure appropriate permissions for initial reservations
Initial reservations are created via the standard do_mmap() in
binfmt_elf, and as a result their permissions are calculated via the
standard helper that calls arch_user_ptr_owning_perms_from_prot().
The latter relies on the current value of PCC to decide whether to
return System and Executive, so we need to ensure that this value is
well-defined when binfmt_elf creates the mappings.

morello_thread_init_user() is called from arch_setup_new_exec(),
which itself is called very early by binfmt_elf. This is therefore a
good place to set PCC to a temporary value with all permissions,
while waiting for the final value set by morello_thread_start().

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>